requirements.txt
#CUDA 12.4 && cuDNN 9.0.1
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
pillow==11.0.0
pandas==2.2.3
